Build scripts? Not actually that clever
Apparently, it's hard to figure out if files were added/deleted to a
directory in a build script. Oh well.
